Meta: Use editable install for development (#123)
* Update install instructions

Submodules have been removed, get rid of their mention in the install
instruction.
Use editable install of faebryk, this needs setuptools >= 64, add it to
requirements.
Show optional usage of venv.

* Remove dependencies from requirements.txt

Since:
 * the package contains the dependencies
 * we do an editable install for development
 * we currently don't fix any specific versions of our dependencies
 we can get rid of them in requirements.txt.

* Get rid of manual adjustment of sys.path

When faebryk is installed, e.g. via an editable install,
adding the path to faebryk manually is not necessary, so get rid of
that.