Fix `apply_netlist` - 2 (#824)
## What?

- Move functionality to apply a design to a PCB directly into the
PCB_Transformer
- Improve robustness for double-sided designs and re-linking
- Re-link using transformer methods, rather than designators

Logging for updates:

![Screenshot 2025-01-11 at 19 39
07](https://github.com/user-attachments/assets/a56a0d54-e4b3-48fc-b487-f0a3916af050)


## Remaining issues:

It appears some caps and resistors aren't being linked properly:

![Screenshot 2025-01-11 at 19 38
51](https://github.com/user-attachments/assets/735a100e-be65-42b4-9235-8c88f0f06c87)


## Later

- Should we keep the netlist infra as is?
- We should be more decisive about the properties we put in the layout.
There's a shitload right now
- We should have a higher fidelity marker. In the schematics, I used a
hash of the DFS content of the PCB (excluding the UUID/hash property
itself). this was great because it meant we could see not only if we
added it, but if someone modified it since then
- We should consider a `--tidy` flag that'll rip up even unmarked things
from the layout. This shouldn't be on be default because it's annoying
to have things like holes removed
- We should either add a netlist target, and potentially make PCB layout
updates optional, which may help us ship a little faster ( @napowderly )

## Fixes:

Fixes: #812
Fixes: #550