Remove cleanup from PCB transformer and buildutil (#852)
It was causing cycles where it'd remove something from the pcb's
"footprints" list, but it'd still be linked via a trait. This means it
wouldn't be output on rebuild, then it'd be re-added on the next run.
This caused cycling where components wouldn't appear every second run
under some circumstances

Fixes: #846