Fix syncing with nested layouts (#880)
Currently, this just supports syncing the layout of the top-level
module.

Implemented nesting modules and synced them as well. However, it
crashed KiCAD and debugging that was impractical and not the best use of
time. I've left the stubs of that work for the future

Additionally, this fixes logging and adds logs to pull, push and reload
groups