Add try except block to get_value in has_simple_value_representation (#905)
Failing build because
`has_simple_value_representation_based_on_params_chain` was expecting
all parameters to provide a value back to construct the 'value' property
in the KiCAD component. In my case, an electrolytic capacitor did not
have a temp-co property, so the whole build would fail.

Now, they get value is wrapped in a try-except block and returns an
empty string for parameters that we fail to get a value for.